+++ EdgeWeightedDigraph.java
@@ -9,7 +9,7 @@ public class EdgeWeightedDigraph extends EdgeWeightedGraph {
   }
 
   @Override
-  public void addEdge(String v, String w, double weight) {
+  public void addEdge(String v, String w, int weight) {
     Edge e = new Edge(v, w, weight);
     addToList(v, e);
     if(!vertices.contains(v)) {
@@ -31,7 +31,7 @@ public class EdgeWeightedDigraph extends EdgeWeightedGraph {
     sb.append("rankdir = LR;" + NEWLINE);
     sb.append("node [shape = circle];" + NEWLINE);
     for (Edge e : getEdges())
-      sb.append(String.format("%s -> %s [label=\"%.3f\"]", e.getV(), e.getW(), e.getWeight()) + NEWLINE);
+      sb.append(String.format("%s -> %s [label=\"%d\"]", e.getV(), e.getW(), e.getWeight()) + NEWLINE);
     sb.append("}" + NEWLINE);
     return sb.toString();
   }
